Add files via upload
The folder "Code" contains the source files.
"sim.sh" is a bash script that creates a folder "sim", executes the program, and saves the output in "sim". 

This version also outputs two files for the occupation vector: one contains all the labels that are active at a given time and the other the total amount of money that corresponds to each label.